# swagger.yaml
openapi: 3.0.3
info:
  title: TaskMate API
  description: |
    REST API для системы управления задачами автосалонов TaskMate.

    ## Аутентификация
    API использует Bearer Token аутентификацию через Laravel Sanctum.
    Получите токен через `POST /api/v1/session` и передавайте его в заголовке:
    ```
    Authorization: Bearer <token>
    ```

    ## Роли пользователей
    - **owner** — владелец, полный доступ ко всем функциям
    - **manager** — менеджер, управление задачами и пользователями в своих автосалонах
    - **observer** — наблюдатель, только просмотр данных
    - **employee** — сотрудник, выполнение назначенных задач

    ## Формат ответов
    Все ответы возвращаются в формате JSON. Успешные ответы содержат поле `data`,
    ошибки содержат поля `message` и `errors`.

    ## Пагинация
    Списки возвращают пагинированные результаты с полями:
    - `data` — массив элементов
    - `meta` — информация о пагинации (current_page, last_page, per_page, total)
    - `links` — ссылки на страницы (first, last, prev, next)
  version: 1.4.0
  contact:
    name: TaskMate Support
servers:
  - url: http://localhost:8007/api
    description: Local Development Server
  - url: https://api.taskmate.ru/api
    description: Production Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer токен, полученный при логине
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        login:
          type: string
          example: "admin"
        full_name:
          type: string
          example: "Иванов Иван Иванович"
        phone_number:
          type: string
          example: "+7 (999) 123-45-67"
          description: Номер телефона (маппинг с поля phone в БД)
        role:
          type: string
          enum: [owner, manager, employee, observer]
          example: "manager"
        dealership_id:
          type: integer
          nullable: true
          example: 1
        dealership:
          type: object
          nullable: true
          description: Основной автосалон (если загружена связь)
          properties:
            id:
              type: integer
            name:
              type: string
            address:
              type: string
            phone:
              type: string
            is_active:
              type: boolean
        dealerships:
          type: array
          description: Все доступные автосалоны (many-to-many)
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              address:
                type: string
              phone:
                type: string
              is_active:
                type: boolean
...
security:
  - bearerAuth: []
paths:
...
  /v1/users/{id}:
    get:
      summary: Получить пользователя
      description: |
        Получить информацию о конкретном пользователе.

        **curl пример:**
        ```bash
        curl -X GET "http://localhost:8007/api/v1/users/5" \
          -H "Authorization: Bearer YOUR_TOKEN"
        ```
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          description: Пользователь не найден

    put:
      summary: Обновить пользователя
      description: |
        Обновить данные пользователя.

        **Доступ:** manager, owner

        **curl пример:**
        ```bash
        curl -X PUT "http://localhost:8007/api/v1/users/5" \
          -H "Authorization: Bearer YOUR_TOKEN" \
          -H "Content-Type: application/json" \
          -d '{
            "full_name": "Петров Пётр Сергеевич",
            "role": "manager"
          }'
        ```
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  minLength: 8
                current_password:
                  type: string
                  description: Требуется при смене своего пароля
                full_name:
                  type: string
                phone:
                  type: string
                role:
                  type: string
                  enum: [owner, manager, employee, observer]
                dealership_id:
                  type: integer
                dealership_ids:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Пользователь обновлён
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '403':
          description: Недостаточно прав
        '404':
          description: Пользователь не найден

    delete:
      summary: Удалить пользователя
      description: |
        Удалить пользователя (soft delete).

        **Доступ:** manager, owner

        **curl пример:**
        ```bash
        curl -X DELETE "http://localhost:8007/api/v1/users/5" \
          -H "Authorization: Bearer YOUR_TOKEN"
        ```
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Пользователь удалён
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Пользователь успешно удалён"
        '403':
          description: Недостаточно прав
        '404':
          description: Пользователь не найден
...